requirements.txt
numpy
pandas
scipy
scikit-learn
statsmodels
pyyaml
pydantic
matplotlib
requests
python-dotenv
rich
